Update Kubernetes logging and monitoring environment user guide

+++ deployment/kubernetes/monitoring/stop_monitoring.sh
@@ -26,9 +26,13 @@ fi
 set +e
 try_command hash kubectl > /dev/null
 
-for i in $(ls $DIR/*.yaml); do
-    kubectl delete -f "$i" &> /dev/null
-done
+if (kubectl get namespace | awk '{print $1}' | grep -q "pgmon"); then
+    kubectl delete -f "$DIR/namespace/namespace.yaml"
+
+    for i in $(find "$DIR" -path "$DIR/namespace" -a -prune -o -name "*.yaml" -print); do
+        kubectl delete -f "$i" &> /dev/null
+    done
+fi
 set -e
 
-rm -rf $DIR/*[0-9].yaml
+echo "Monitoring are stopping..."
